+++ benchmark/c_str-functions/c_str.h
