+++ benchmark/c_str-functions/corpus.h
@@ -1,4 +1,4 @@
-#include "atoi.h"
+#include "c_str.h"
 #include "zoo/pp/platform.h"
 
 #include <vector>
